backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
pymongo==4.6.0
pinecone-client==2.2.4
openai==1.3.5
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2